messaging-demo/issue-revoke-cat/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Deprecated==1.2.13
idna==3.4
jwcrypto==1.4.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
wrapt==1.14.1
